removed, strip stupid windows types out. we are c++11 and god gave us good data types for a reason

+++ xbmc/filesystem/NFSDirectory.cpp
@@ -25,6 +25,7 @@
 
 #include "NFSDirectory.h"
 #include "FileItem.h"
+#include "linux/ConvUtils.h"
 #include "utils/log.h"
 #include "utils/URIUtils.h"
 #include "utils/StringUtils.h"
@@ -251,9 +252,9 @@ bool CNFSDirectory::GetDirectory(const CURL& url, CFileItemList &items)
         lTimeDate = nfsdirent->ctime.tv_sec;
       }
 
-      LONGLONG ll = Int32x32To64(lTimeDate & 0xffffffff, 10000000) + 116444736000000000ll;
-      fileTime.dwLowDateTime = (DWORD) (ll & 0xffffffff);
-      fileTime.dwHighDateTime = (DWORD)(ll >> 32);
+      long long ll = Int32x32To64(lTimeDate & 0xffffffff, 10000000) + 116444736000000000ll;
+      fileTime.dwLowDateTime = (uint32_t) (ll & 0xffffffff);
+      fileTime.dwHighDateTime = (uint32_t)(ll >> 32);
       FileTimeToLocalFileTime(&fileTime, &localTime);
 
       CFileItemPtr pItem(new CFileItem(nfsdirent->name));
